Move a lot of terminalService to common/browser
